Removed undeploy when a swadl file changed.
Now that the workflow id is required, when a swadl file is changed, the previous deployed version is not used anymore for new executions but only for pending ones. For this reason, WorkflowFolderWatcher does not need anymore to undeploy a workflow when the file changed. Calling undeploy method is still required when a swadl file is deleted.